Improved output for user
Changed the rsync command in the script to eliminate re-direction. The logfile is set using the option --log-file= and progress of the rsync command is reported to the user via the option --progress. Tested and found to work on a MacBook with Sonoma 14.6.1 and the --log-file format by default is an ASCII text file.